+++ src/pyqt_plugins/_version.py
