+++ src/pyqt_plugins/badplugin/badwidgetplugin.py
