+++ src/pyqt_plugins/examplebutton.py
