+++ src/pyqt_plugins/examples/exampleqmlitem.py
