+++ src/pyqt_plugins/examples/exampleqmlitemplugin.py
