+++ src/pyqt_plugins/examples/qmlapp.qml
