+++ src/pyqt_plugins/examples/qmldir
