+++ src/pyqt_plugins/examples/qmltest.qml
