+++ src/pyqt_plugins/tests/test_plugins.py
