+++ src/pyqt_plugins/tests/testbuttonplugin.py
